feat(logout-prompt): Reintegrated logout prompt usage as a separate component
