+++ frontend/src/App.jsx
@@ -2,7 +2,9 @@ import { useContext } from "react";
 
 import Login from "./components/Login"
 import CurrentSongOffCanvas from "./components/CurrentSongOffCanvas"
+import LogoutPrompt from "./components/LogoutPrompt";
 import Main from "./components/Main-window"
+
 import { Context } from "./contexts/Context"
 
 import "./App.css"
@@ -13,6 +15,7 @@ function App() {
   return <>
     {accessToken ? <Main /> : <Login />}
     <CurrentSongOffCanvas />
+    <LogoutPrompt />
   </>;
 }
